Grand unification of init and finish functions
This unifies the init/finish functions for the various ciphers into
canonical functions, cuda_device_init and cuda_device_finish, defined in
cuda_common.cu.

cuda_device_init simply allocs memory on the device and stores memory to
it in device_data, which has to be passed to the single crypt-functions
for the ciphers. The same goes for host_data.